Migrated component page to typescript
