--- src/views/components/shared/ComponentPage/index.js
@@ -1,1 +1,0 @@
-export { default } from './ComponentPage';
