+++ src/views/components/shared/ComponentPage/index.ts
@@ -1,0 +1,1 @@
+export * from './ComponentPage';
